+++ data/src/main/java/com/semicolon/data/remote/response/challenge/ChallengeDetailResponse.kt
@@ -4,15 +4,15 @@ import com.google.gson.annotations.SerializedName
 import java.time.LocalDateTime
 
 data class ChallengeDetailResponse(
-    val name: String,
-    val content: String,
-    val goal: Int,
-    val award: String,
+    @SerializedName("name") val name: String,
+    @SerializedName("content") val content: String,
+    @SerializedName("goal") val goal: Int,
+    @SerializedName("award") val award: String,
     @SerializedName("image_url") val imageUrl: String,
     @SerializedName("start_at") val startAt: LocalDateTime,
     @SerializedName("end_at") val endAt: LocalDateTime,
-    val scope: String,
+    @SerializedName("scope") val scope: String,
     @SerializedName("participant_count") val participantCount: Int,
-    val isMine: Boolean,
-    val writer: ChallengeWriterResponse
+    @SerializedName("isMine") val isMine: Boolean,
+    @SerializedName("writer") val writer: ChallengeWriterResponse
 )
