+++ data/src/main/java/com/semicolon/data/remote/response/challenge/ChallengeWriterResponse.kt
@@ -3,7 +3,7 @@ package com.semicolon.data.remote.response.challenge
 import com.google.gson.annotations.SerializedName
 
 data class ChallengeWriterResponse(
-    val id: Long,
-    val name: String,
+    @SerializedName("id") val id: Long,
+    @SerializedName("name") val name: String,
     @SerializedName("profile_image_url") val profileUrl: String
 )
